I displayed the forecast on the interface using data from the API.
Stored the response from forecast API and made it loop through the forecast daily array of 7 items, then got the index of the current item in the array. Retrieved the day using timestamp in the function which returns days, then set the icon using the icon URL, rounded the maximum and minimum temperature and lastly checked if index is greater than 5 ,it should not display the day only if its less it should show the first 6 days.